class/week3/wee3_project/requirements.txt
markupsafe==2.0.1
check50
pillow

